:recycle: Isolate device input codes to `inputCodes.ts`
