+++ src/constants/index.ts
@@ -1,191 +1,3 @@
-const mouseCodes = {
-    Mouse1: 0x0,
-    Mouse2: 0x1,
-    Mouse3: 0x2,
-    Mouse4: 0x3,
-    Mouse5: 0x4,
-    Mouse6: 0x5,
-    Mouse7: 0x6,
-    Mouse8: 0x7,
-    MouseWheelUp: 0x8,
-    MouseWheelDown: 0x9,
-    MouseMove: 0xa,
-};
-
-const gamepadCodes = {
-    Up: 0x0001,
-    Down: 0x0002,
-    Left: 0x0004,
-    Right: 0x008,
-    "360_Start": 0x0010,
-    "360_Back": 0x0020,
-    "360_L3": 0x0040,
-    "360_R3": 0x0080,
-    "360_LB": 0x0100,
-    "360_RB": 0x0200,
-    "360_A": 0x1000,
-    "360_B": 0x2000,
-    "360_X": 0x4000,
-    "360_Y": 0x8000,
-    "360_LT": 0x0009,
-    "360_RT": 0x000a,
-    "360_LS": 0x000b,
-    "360_RS": 0x000c,
-};
-
-const keyboardCodes = {
-    Esc: 0x01,
-    "1": 0x02,
-    "2": 0x03,
-    "3": 0x04,
-    "4": 0x05,
-    "5": 0x06,
-    "6": 0x07,
-    "7": 0x08,
-    "8": 0x09,
-    "9": 0x0a,
-    "0": 0x0b,
-    Hyphen: 0x0c,
-    Equal: 0x0d,
-    Backspace: 0x0e,
-    Tab: 0x0f,
-    Q: 0x10,
-    W: 0x11,
-    E: 0x12,
-    R: 0x13,
-    T: 0x14,
-    Y: 0x15,
-    U: 0x16,
-    I: 0x17,
-    O: 0x18,
-    P: 0x19,
-    Bracketleft: 0x1a,
-    Bracketright: 0x1b,
-    Enter: 0x1c,
-    "L-Ctrl": 0x1d,
-    A: 0x1e,
-    S: 0x1f,
-    D: 0x20,
-    F: 0x21,
-    G: 0x22,
-    H: 0x23,
-    J: 0x24,
-    K: 0x25,
-    L: 0x26,
-    Semicolon: 0x27,
-    Quotesingle: 0x28,
-    Tilde: 0x29,
-    "L-Shift": 0x2a,
-    Backslash: 0x2b,
-    Z: 0x2c,
-    X: 0x2d,
-    C: 0x2e,
-    V: 0x2f,
-    B: 0x30,
-    N: 0x31,
-    M: 0x32,
-    Comma: 0x33,
-    Period: 0x34,
-    Slash: 0x35,
-    "R-Shift": 0x36,
-    NumPadMult: 0x37,
-    "L-Alt": 0x38,
-    Space: 0x39,
-    CapsLock: 0x3a,
-    F1: 0x3b,
-    F2: 0x3c,
-    F3: 0x3d,
-    F4: 0x3e,
-    F5: 0x3f,
-    F6: 0x40,
-    F7: 0x41,
-    F8: 0x42,
-    F9: 0x43,
-    F10: 0x44,
-    NumLock: 0x45,
-    ScrollLock: 0x46,
-    NumPad7: 0x47,
-    NumPad8: 0x48,
-    NumPad9: 0x49,
-    NumPadMinus: 0x4a,
-    NumPad4: 0x4b,
-    NumPad5: 0x4c,
-    NumPad6: 0x4d,
-    NumPadPlus: 0x4e,
-    NumPad1: 0x4f,
-    NumPad2: 0x50,
-    NumPad3: 0x51,
-    NumPad0: 0x52,
-    NumPadDec: 0x53,
-    DIK_OEM_102: 0x56,
-    F11: 0x57,
-    F12: 0x58,
-    F13: 0x64,
-    F14: 0x65,
-    F15: 0x66,
-    DIK_KANA: 0x70,
-    DIK_ABNT_C1: 0x73,
-    DIK_CONVERT: 0x79,
-    DIK_NOCONVERT: 0x7b,
-    DIK_OEM_7D: 0x7d,
-    DIK_ABNT_C2: 0x7e,
-    NumPadEqual: 0x8d,
-    "Prev Track": 0x90,
-    DIK_AT: 0x91,
-    ":": 0x92,
-    DIK_UNDERLINE: 0x93,
-    DIK_KANJI: 0x94,
-    DIK_STOP: 0x95,
-    DIK_AX: 0x96,
-    DIK_UNLABELED: 0x97,
-    "Next Track": 0x99,
-    NumPadEnter: 0x9c,
-    "R-Ctrl": 0x9d,
-    Mute: 0xa0,
-    Calc: 0xa1,
-    "Play/Pause": 0xa2,
-    MediaStop: 0xa4,
-    "Vol-": 0xae,
-    "Vol+": 0xb0,
-    WebHome: 0xb2,
-    NumPadComma: 0xb3,
-    NumPadDivide: 0xb5,
-    PrintSrc: 0xb7,
-    "R-Alt": 0xb8,
-    Pause: 0xc5,
-    Home: 0xc7,
-    Up: 0xc8,
-    PgUp: 0xc9,
-    Left: 0xcb,
-    Right: 0xcd,
-    End: 0xcf,
-    Down: 0xd0,
-    PgDn: 0xd1,
-    Insert: 0xd2,
-    Delete: 0xd3,
-    "L-Windows": 0xdb,
-    "R-Windows": 0xdc,
-    Apps: 0xdd,
-    Power: 0xde,
-    Sleep: 0xdf,
-    Wake: 0xe3,
-    WebSearch: 0xe5,
-    WebFavorites: 0xe6,
-    WebRefresh: 0xe7,
-    WebStop: 0xe8,
-    WebForward: 0xe9,
-    WebBack: 0xea,
-    "My Computer": 0xeb,
-    Mail: 0xec,
-    MediaSelect: 0xed,
-};
-
-export const codes = {
-    keyboard: keyboardCodes,
-    mouse: mouseCodes,
-    gamepad: gamepadCodes,
-};
-
 export const STANDARD_FILE_HEADER = `// 1st field: User event name.  DO NOT ALTER!  This field is used to ID events in the code							
 // 2nd: Keyboard key ID that will proc this event.  A value of 0xff means the event is unmapped for this device.							
 // 3rd: Mouse button ID that will proc this event.
